Upload maven file instead of downloading from maven

+++ maven.sh
@@ -1,16 +1,18 @@
 #!/bin/bash
 
-# Define the desired Maven version
-maven_version=$1
+if [ ! -f "tmp/apache-maven.tar.gz" ]; then
+  # Define the desired Maven version
+  maven_version=$1
 
-# Define the download URL
-download_url="https://apache.osuosl.org/maven/maven-3/${maven_version}/binaries/apache-maven-${maven_version}-bin.tar.gz"
+  # Define the download URL
+  download_url="https://apache.osuosl.org/maven/maven-3/${maven_version}/binaries/apache-maven-${maven_version}-bin.tar.gz"
 
-# Download Maven
-echo Downloading...$download_url
-mkdir tmp
-curl -fsSL "$download_url" -o tmp/apache-maven.tar.gz
-#wget "$download_url" tmp/apache-maven.tar.gz
+  # Download Maven
+  echo Downloading...$download_url
+  mkdir tmp
+  curl -fsSL "$download_url" -o tmp/apache-maven.tar.gz
+  #wget "$download_url" tmp/apache-maven.tar.gz
+fi
 
 # Extract Maven to a temporary directory
 tar xf tmp/apache-maven.tar.gz -C tmp
